+++ src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import {InitialNavigation} from '@angular/router';
+import type {InitialNavigation} from '@angular/router';
 
 export const environment = {
   production: true,
@@ -11,5 +11,6 @@ export const environment = {
     appId: '1:665830225099:web:18e0669d5847a4b047974e',
     measurementId: 'G-1LXY5W5Z9H',
   },
+  reCAPTCHAKey: '6Le6rbwoAAAAABOHJ80tHDRyHEZCo9Hmi-MC-P1Z',
   initialNavigation: 'enabledBlocking' as InitialNavigation,
 };