+++ src/environments/environment.test.ts
@@ -2,7 +2,7 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-import {InitialNavigation} from '@angular/router';
+import type {InitialNavigation} from '@angular/router';
 
 export const environment = {
   production: false,
@@ -15,6 +15,7 @@ export const environment = {
     appId: '1:665830225099:web:18e0669d5847a4b047974e',
     measurementId: null,
   },
+  reCAPTCHAKey: '',
   initialNavigation: 'enabledNonBlocking' as InitialNavigation,
 };